Corrected Detection Error
With the changes from the last commit, any second+ scan type was reporting detection because it was looking for a present scan log to confirm detection.
Instead, we're now setting the DETECTIONTYPE_DETECTION (where the word DETECTIONTYPE changes based on the scan type) variable equal to 1 when a detection has occurred to trigger the correct text output.