Arbitrary Code Injection fix
Found an issue with arbitrary code injection into the malscan.conf file, which would then be run when the conf file was sourced.
Mitigation: Set the `malscan -s` flag to only be usable by root.
This is a best practice anyway, but makes it so only the root user can exploit this. Since they have unfettered command line access anyway, there's no issue.
This updates #12 though i'm going to be looking at better ways of getting the variables from the conf file